getDuration() now overridden to return hlsjs reported duration
Also timeupdate event now always contains the actual time instead of
trying to round up to the duration when dvr wasn't being used
(incorrectly). I think the reason this was there was to make the seek
bar stick to the right when dvr wasn't being used but this might make
more sense being handled in MediaControl so that the correct values are
still reported in the event.

TODO:
- update MediaControl so that it renders the seek bar slider at the far
right when dvr is disabled.
- Figure out what do do in getPlayableStartTime() (See comment)